# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Lingow Account, Usage and Delivery API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /auth/anonymous:
    post:
      operationId: createAnonymousAccount
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResult' } } } }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /auth/verification-codes:
    post:
      operationId: createPhoneChallenge
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PhoneChallengeRequest' } } } }
      responses:
        '202': { description: Accepted, content: { application/json: { schema: { type: object, required: [challenge_id], properties: { challenge_id: { type: string } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /auth/phone/login:
    post:
      operationId: verifyPhone
      security:
        - {}
        - accountContext: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PhoneVerifyRequest' } } } }
      responses:
        '200': { description: Verified, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResult' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /auth/token/refresh:
    post:
      operationId: refreshToken
      requestBody: { $ref: '#/components/requestBodies/RefreshToken' }
      responses:
        '200': { description: Refreshed, content: { application/json: { schema: { $ref: '#/components/schemas/Tokens' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /auth/logout:
    post:
      operationId: logout
      requestBody: { $ref: '#/components/requestBodies/RefreshToken' }
      responses:
        '204': { description: Logged out }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /account/me:
    get:
      operationId: getMe
      security: [{ accountContext: [] }]
      responses:
        '200': { description: Current account, content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /voice-sessions/{id}/usage:
    get:
      operationId: getSessionUsage
      security: [{ accountContext: [] }]
      parameters: [{ $ref: '#/components/parameters/SessionID' }]
      responses:
        '200': { description: Session usage, content: { application/json: { schema: { $ref: '#/components/schemas/SessionUsageSummary' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /usage/summary:
    get:
      operationId: getAccountUsage
      security: [{ accountContext: [] }]
      parameters:
        - { name: period_start, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: period_end, in: query, required: true, schema: { type: string, format: date-time } }
      responses:
        '200': { description: Account usage, content: { application/json: { schema: { $ref: '#/components/schemas/AccountUsageSummary' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /outbound-messages:
    post:
      operationId: createOutboundMessage
      security: [{ accountContext: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateMessageRequest' } } } }
      responses:
        '202': { description: Queued, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /outbound-messages/{message_id}:
    get:
      operationId: getOutboundMessage
      security: [{ accountContext: [] }]
      parameters: [{ $ref: '#/components/parameters/MessageID' }]
      responses:
        '200': { description: Message, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /outbound-deliveries/{message_id}/retry:
    post:
      operationId: retryOutboundMessage
      security: [{ accountContext: [] }]
      parameters:
        - { $ref: '#/components/parameters/MessageID' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202': { description: Retry queued, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /account/message-preferences:
    get:
      operationId: listMessagePreferences
      security: [{ accountContext: [] }]
      responses:
        '200': { description: Preferences, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Preference' } } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /account/message-preferences/{channel}:
    put:
      operationId: putMessagePreference
      security: [{ accountContext: [] }]
      parameters: [{ name: channel, in: path, required: true, schema: { $ref: '#/components/schemas/Channel' } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false, required: [enabled], properties: { enabled: { type: boolean } } } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Preference' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /account/wecom-bots/{destination_ref}:
    put:
      operationId: configureWeComBot
      security: [{ accountContext: [] }]
      parameters: [{ name: destination_ref, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 100 } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [webhook_url]
              properties:
                webhook_url: { type: string, format: uri, description: Enterprise WeChat group-bot webhook. Stored encrypted and never returned. }
      responses:
        '204': { description: Webhook verified and stored for the authenticated account. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
components:
  securitySchemes:
    accountContext:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token validated by authentication middleware; account_id comes from the trusted request context.
  parameters:
    SessionID: { name: id, in: path, required: true, schema: { type: string } }
    MessageID: { name: message_id, in: path, required: true, schema: { type: string } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 200 } }
  requestBodies:
    RefreshToken:
      required: true
      content: { application/json: { schema: { type: object, additionalProperties: false, required: [refresh_token], properties: { refresh_token: { type: string, minLength: 1 } } } } }
  responses:
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: Missing or invalid account context, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: Authenticated account does not own the requested resource, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Requested resource was not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Conflict: { description: Request conflicts with the current resource state or existing data, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    TooManyRequests: { description: Phone verification requests are temporarily rate limited, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotImplemented: { description: Business implementation is intentionally pending, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    RealtimeRuntimeState:
      type: string
      enum: [stopped, starting, listening, asr_processing, translating, tts_processing, playing, stopping, failed]
    RealtimeRuntimeErrorCode:
      type: string
      enum: [realtime_start_failed, realtime_stop_failed]
    RealtimeRuntimeSnapshot:
      type: object
      additionalProperties: false
      required: [session_id, runtime_state, current_turn_id, current_playback_id, last_error_code, updated_at]
      properties:
        session_id: { type: string }
        runtime_state: { $ref: '#/components/schemas/RealtimeRuntimeState' }
        current_turn_id: { type: [string, 'null'] }
        current_playback_id: { type: [string, 'null'] }
        last_error_code: { type: [string, 'null'] }
        updated_at: { type: string, format: date-time }
    WebRTCConnectionState:
      type: string
      enum: [new, connecting, connected, disconnected, failed, closed]
    WebRTCErrorCode:
      type: string
      enum: [webrtc_connection_not_found, webrtc_connection_not_ready, webrtc_connection_failed]
    WebRTCConnectionSnapshot:
      type: object
      additionalProperties: false
      required: [session_id, connection_id, state, version, updated_at]
      properties:
        session_id: { type: string }
        connection_id: { type: string }
        state: { $ref: '#/components/schemas/WebRTCConnectionState' }
        version: { type: integer, format: int64, minimum: 1 }
        updated_at: { type: string, format: date-time }
    Account:
      type: object
      required: [id, kind, created_at]
      properties: { id: { type: string }, kind: { type: string, enum: [anonymous, registered] }, created_at: { type: string, format: date-time } }
    Tokens:
      type: object
      required: [access_token, refresh_token, expires_at]
      properties: { access_token: { type: string }, refresh_token: { type: string }, expires_at: { type: string, format: date-time } }
    AuthResult:
      type: object
      required: [account, tokens]
      properties: { account: { $ref: '#/components/schemas/Account' }, tokens: { $ref: '#/components/schemas/Tokens' } }
    PhoneChallengeRequest:
      type: object
      additionalProperties: false
      required: [phone]
      properties:
        phone:
          type: string
          pattern: '^\+[1-9][0-9]{7,14}$'
          description: Canonical E.164 phone number without spaces or separators.
    PhoneVerifyRequest:
      type: object
      additionalProperties: false
      required: [challenge_id, code]
      properties:
        challenge_id: { type: string, minLength: 1 }
        code: { type: string, pattern: '^[0-9]{6}$' }
        anonymous_account_id:
          type: string
          description: When supplied, the request must include a Bearer access token for this same anonymous account.
    Stage: { type: string, enum: [asr, translation, tts, diarization] }
    SessionUsageSummary:
      type: object
      required: [account_id, session_id, totals]
      properties:
        account_id: { type: string }
        session_id: { type: string }
        totals:
          type: array
          items: { $ref: '#/components/schemas/UsageStageTotal' }
    AccountUsageSummary:
      type: object
      required: [account_id, period_start, period_end, totals]
      properties:
        account_id: { type: string }
        period_start: { type: string, format: date-time }
        period_end: { type: string, format: date-time }
        totals:
          type: array
          items: { $ref: '#/components/schemas/UsageStageTotal' }
    UsageStageTotal:
      type: object
      required: [service_type, input_tokens, output_tokens, audio_duration_ms, cost_amount, currency]
      oneOf:
        - properties:
            cost_amount: { const: '' }
            currency: { const: '' }
        - properties:
            cost_amount: { type: string, pattern: '^(?:0|[1-9][0-9]{0,11})(?:\.[0-9]{1,8})?$' }
            currency: { type: string, pattern: '^[A-Z]{3}$' }
      properties:
        service_type: { $ref: '#/components/schemas/Stage' }
        input_tokens: { type: integer, format: int64, minimum: 0 }
        output_tokens: { type: integer, format: int64, minimum: 0 }
        audio_duration_ms: { type: integer, format: int64, minimum: 0 }
        cost_amount: { type: string, description: Non-negative NUMERIC(20,8)-compatible amount, or an empty string when pricing is unavailable. }
        currency: { type: string, description: ISO 4217 currency paired with cost_amount, or an empty string when pricing is unavailable. }
    Channel: { type: string, enum: [email, wecom_bot] }
    CreateMessageRequest:
      type: object
      additionalProperties: false
      required: [channel, destination_ref, turn_ids]
      properties: { channel: { $ref: '#/components/schemas/Channel' }, destination_ref: { type: string }, turn_ids: { type: array, minItems: 1, uniqueItems: true, items: { type: string } } }
    FinalTurnSnapshot:
      type: object
      required: [turn_id, session_id, participant_id, speaker_label_snapshot, source_language, target_language, language_config_version, source_text, translated_text, created_at]
      properties:
        turn_id: { type: string }
        session_id: { type: string }
        participant_id: { type: [string, 'null'] }
        speaker_label_snapshot: { type: [string, 'null'] }
        source_language: { type: string }
        target_language: { type: string }
        language_config_version: { type: integer, format: int64, minimum: 1 }
        source_text: { type: string }
        translated_text: { type: string }
        created_at: { type: string, format: date-time }
    Message:
      type: object
      required: [id, account_id, channel, destination_ref, snapshot_version, turns, status, attempts, last_error_code, created_at, updated_at]
      properties:
        id: { type: string }
        account_id: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        destination_ref: { type: string }
        snapshot_version: { type: integer, minimum: 1 }
        turns: { type: array, items: { $ref: '#/components/schemas/FinalTurnSnapshot' } }
        status: { type: string, enum: [queued, sending, sent, failed, retrying, cancelled] }
        attempts: { type: integer, minimum: 0 }
        last_error_code: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Preference:
      type: object
      required: [account_id, channel, enabled, verified, updated_at]
      properties: { account_id: { type: string }, channel: { $ref: '#/components/schemas/Channel' }, enabled: { type: boolean }, verified: { type: boolean }, updated_at: { type: string, format: date-time } }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id, retryable, details]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            retryable: { type: boolean }
            details: { type: object, additionalProperties: true }